requirements.txt
streamlit~=1.22.0
Pillow~=9.5.0
rembg~=2.0.35
requests~=2.30.0
numpy~=1.24.3